requirements.txt
discord.py
wavelink
py-cord